Returning Uint8Array instead of Float32Array

+++ src/BabylonCpp/include/babylon/misc/dds.h
@@ -143,12 +143,12 @@ private:
   static Float32Array _GetFloatRGBAArrayBuffer(float width, float height, int dataOffset,
                                                size_t dataLength, const Uint8Array& arrayBuffer,
                                                int lod);
-  static Float32Array _GetFloatAsUIntRGBAArrayBuffer(float width, float height, int dataOffset,
-                                                     size_t dataLength,
-                                                     const Uint8Array& arrayBuffer, int lod);
-  static Float32Array _GetHalfFloatAsUIntRGBAArrayBuffer(float width, float height, int dataOffset,
-                                                         size_t dataLength,
-                                                         const Uint8Array& arrayBuffer, int lod);
+  static Uint8Array _GetFloatAsUIntRGBAArrayBuffer(float width, float height, int dataOffset,
+                                                   size_t dataLength, const Uint8Array& arrayBuffer,
+                                                   int lod);
+  static Uint8Array _GetHalfFloatAsUIntRGBAArrayBuffer(float width, float height, int dataOffset,
+                                                       size_t dataLength,
+                                                       const Uint8Array& arrayBuffer, int lod);
   static Uint8Array _GetRGBAArrayBuffer(float width, float height, int dataOffset,
                                         size_t dataLength, const Uint8Array& arrayBuffer,
                                         int rOffset, int gOffset, int bOffset, int aOffset);
